requirements.txt
flask
pymongo
requests
python-dotenv
apscheduler
pandas
matplotlib
